src/requirements.txt
requests
opencensus-ext-azure
opencensus-ext-requests